Exceptions + Product mapper + some fixes

+++ src/main/java/com/kodilla/ecommercee/domain/dto/ProductDto.java
@@ -3,12 +3,14 @@ package com.kodilla.ecommercee.domain.dto;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
+import java.math.BigDecimal;
+
 @Getter
 @AllArgsConstructor
 public class ProductDto {
-    private int id;
+    private long id;
     private String name;
     private String description;
-    private int price;
+    private BigDecimal price;
     private String groupId;
 }
